Moved `enable_rhrepo_and_fetchid` function to api.utils module.
This helps both test_repository module and test_smoke module
to use it.

api.utils is choosen as the code is very specific to the API.
